Remove GlobalMemoryStore
