+++ sdk/src/main/java/com/oursky/authgear/data/token/RefreshTokenRepo.kt
@@ -1,0 +1,7 @@
+package com.oursky.authgear.data.token
+
+internal interface RefreshTokenRepo {
+    fun setRefreshToken(namespace: String, refreshToken: String)
+    fun getRefreshToken(namespace: String): String?
+    fun deleteRefreshToken(namespace: String)
+}
