+++ sdk/src/main/java/com/oursky/authgear/data/token/TokenRepo.kt
@@ -1,10 +1,6 @@
 package com.oursky.authgear.data.token
 
 internal interface TokenRepo {
-    fun setRefreshToken(namespace: String, refreshToken: String)
-    fun getRefreshToken(namespace: String): String?
-    fun deleteRefreshToken(namespace: String)
-
     fun setOIDCCodeVerifier(namespace: String, verifier: String)
     fun getOIDCCodeVerifier(namespace: String): String?
